requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
anyio==4.12.1
attrs==25.4.0
dotenv==0.9.9
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
grpcio==1.78.0
grpcio-tools==1.78.0
idna==3.11
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
multidict==6.7.1
numpy==2.4.2
propcache==0.4.1
protobuf==6.33.5
py-cord==2.7.0
python-dotenv==1.2.1
setuptools==82.0.0
sniffio==1.3.1
typing_extensions==4.15.0
valkey-glide==2.2.6
yarl==1.22.0
